requirements.txt
# Temel kütüphaneler
pandas>=1.3.0
numpy>=1.21.0

# PDF ve doküman işleme
fpdf2>=2.7.4
python-docx>=0.8.11
PyPDF2>=2.0.0
pdfminer.six>=20221105

# Veri görselleştirme
matplotlib>=3.5.0
seaborn>=0.11.2
plotly>=5.10.0  # İnteraktif grafikler

# Web ve veri çekme
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
selenium>=4.7.0  # Dinamik web sayfaları
yfinance

# Streamlit ve arayüz
streamlit>=1.12.0
streamlit-aggrid>=0.3.0
streamlit-extras>=0.2.0

# Veri işleme
openpyxl>=3.0.0
python-dateutil>=2.8.2
scipy>=1.7.0
scikit-learn>=1.0.0
xlrd>=2.0.0

# Diğer
typing-extensions>=4.0.0
loguru>=0.6.0
numba>=0.56.0
cryptography>=38.0.0